requirements.txt
pytest
pytest-playwright
playwright
allure-pytest
ImageGrab
io
AttachmentType